Fix build warnings

+++ src/exchange/exch_coinbase_feed_callbacks.cpp
@@ -7,54 +7,66 @@
 using namespace slick::sim::exch;
 using namespace slick::sim;
 
-namespace {
+namespace
+{
     auto &sym_mgr = SymbolManager::instance();
 }
 
-void CoinbaseExchange::onMarketDataConnected(WebSocketClient* client) {
-    LOG_INFO("WebSocket {:p} market data connected", (void*)client);
+void CoinbaseExchange::onMarketDataConnected(WebSocketClient *client)
+{
+    LOG_INFO("WebSocket {:p} market data connected", (void *)client);
 }
 
-void CoinbaseExchange::onMarketDataDisconnected(WebSocketClient* client) {
-    LOG_INFO("WebSocket {:p} market data disconnected", (void*)client);
+void CoinbaseExchange::onMarketDataDisconnected(WebSocketClient *client)
+{
+    LOG_INFO("WebSocket {:p} market data disconnected", (void *)client);
 
     // Find the feed that owns this client
     std::shared_ptr<md_feed::MDFeed> disconnected_feed;
-    for (auto& feed : md_feeds_) {
-        auto* live_feed = dynamic_cast<md_feed::CoinbaseLiveWSFeed*>(feed.get());
-        if (live_feed && live_feed->getClient() == client) {
+    for (auto &feed : md_feeds_)
+    {
+        auto *live_feed = dynamic_cast<md_feed::CoinbaseLiveWSFeed *>(feed.get());
+        if (live_feed && live_feed->getClient() == client)
+        {
             disconnected_feed = feed;
             break;
         }
     }
 
-    if (!disconnected_feed) {
-        LOG_WARN("WebSocket {:p} disconnected but no matching feed found", (void*)client);
+    if (!disconnected_feed)
+    {
+        LOG_WARN("WebSocket {:p} disconnected but no matching feed found", (void *)client);
         return;
     }
 
     // Collect symbols associated with this feed
     std::vector<std::string> symbols;
-    for (auto& [sym, feed] : map_symbol_feed_) {
-        if (feed == disconnected_feed) {
+    for (auto &[sym, feed] : map_symbol_feed_)
+    {
+        if (feed == disconnected_feed)
+        {
             symbols.push_back(sym);
         }
     }
 
     // Remove old feed from tracking
     md_feeds_.erase(std::remove(md_feeds_.begin(), md_feeds_.end(), disconnected_feed), md_feeds_.end());
-    for (const auto& sym : symbols) {
+    for (const auto &sym : symbols)
+    {
         map_symbol_feed_.erase(sym);
     }
 
-    if (symbols.empty()) {
+    if (symbols.empty())
+    {
         return;
     }
 
     // Mark symbols as pending subscription so the snapshot is re-requested on reconnect
-    for (const auto& sym : symbols) {
-        auto* symbol = sym_mgr.getSymbol(sym);
-        if (symbol) {
+    for (const auto &sym : symbols)
+    {
+        auto *symbol = sym_mgr.getSymbol(sym);
+        if (symbol)
+        {
             pending_md_subscription_[coinbase::WebSocketChannel::LEVEL2].emplace(symbol);
             pending_md_subscription_[coinbase::WebSocketChannel::MARKET_TRADES].emplace(symbol);
             trade_snapshots_.erase(symbol);
@@ -64,25 +76,30 @@ void CoinbaseExchange::onMarketDataDisconnected(WebSocketClient* client) {
     // Create and start a new feed
     auto new_feed = std::make_shared<md_feed::CoinbaseLiveWSFeed>(ws_mux_, this, symbols);
     md_feeds_.emplace_back(new_feed);
-    for (const auto& sym : symbols) {
+    for (const auto &sym : symbols)
+    {
         map_symbol_feed_.emplace(sym, new_feed);
     }
     new_feed->start();
 
-    LOG_INFO("WebSocket {:p} restarted market data feed for {} symbol(s)", (void*)client, symbols.size());
+    LOG_INFO("WebSocket {:p} restarted market data feed for {} symbol(s)", (void *)client, symbols.size());
 }
 
-void CoinbaseExchange::onUserDataConnected(WebSocketClient* client) {
-    LOG_INFO("WebSocket {:p} user data connected", (void*)client);
+void CoinbaseExchange::onUserDataConnected(WebSocketClient *client)
+{
+    LOG_INFO("WebSocket {:p} user data connected", (void *)client);
 }
 
-void CoinbaseExchange::onUserDataDisconnected(WebSocketClient* client) {
-    LOG_INFO("WebSocket {:p} user data disonnected", (void*)client);
+void CoinbaseExchange::onUserDataDisconnected(WebSocketClient *client)
+{
+    LOG_INFO("WebSocket {:p} user data disonnected", (void *)client);
 }
 
-void CoinbaseExchange::onLevel2Snapshot(WebSocketClient* /* client */, uint64_t seq_num, const coinbase::Level2UpdateBatch& snapshot) {
+void CoinbaseExchange::onLevel2Snapshot(WebSocketClient * /* client */, uint64_t seq_num, const coinbase::Level2UpdateBatch &snapshot)
+{
     auto *symbol = sym_mgr.getSymbol(snapshot.product_id);
-    if (!symbol) [[unlikely]] {
+    if (!symbol) [[unlikely]]
+    {
         return;
     }
 
@@ -92,7 +109,8 @@ void CoinbaseExchange::onLevel2Snapshot(WebSocketClient* /* client */, uint64_t 
     std::vector<MDLevel> level_updates;
     std::vector<MDTrade> trade_updates;
 
-    for (auto it_update = snapshot.updates.rbegin(), it_last = std::prev(snapshot.updates.rend()); it_update != snapshot.updates.rend(); ++it_update) {
+    for (auto it_update = snapshot.updates.rbegin(), it_last = std::prev(snapshot.updates.rend()); it_update != snapshot.updates.rend(); ++it_update)
+    {
         auto &update = *it_update;
         auto price = to_price_t(update.price_level);
         auto side = static_cast<Side>(update.side);
@@ -101,17 +119,21 @@ void CoinbaseExchange::onLevel2Snapshot(WebSocketClient* /* client */, uint64_t 
 
         auto order_id = utils::nextOrderId();
         auto trade_summaries = symbol->matching_engine_->match(side, order_id, price, qty, *symbol->order_book_.get(), update.event_time, seq_num);
-        if (qty) {
+        if (qty)
+        {
             symbol->order_book_->addOrder(order_id, book_side, price, qty, update.event_time, seq_num, it_update == it_last);
         }
 
-        if (!trade_summaries.empty()) {
-            for (const auto &summary : trade_summaries) {
+        if (!trade_summaries.empty())
+        {
+            for (const auto &summary : trade_summaries)
+            {
                 publishTradeSummary(symbol->symbol_.c_str(), summary);
             }
         }
 
-        if (it == pending_l2_subscriptions.end()) {
+        if (it == pending_l2_subscriptions.end())
+        {
             level_update_buffer_.emplace_back(MDLevel{
                 .event_time = update.event_time,
                 .seq_num = seq_num,
@@ -119,23 +141,25 @@ void CoinbaseExchange::onLevel2Snapshot(WebSocketClient* /* client */, uint64_t 
                 .qty = qty,
                 .num_orders = 1,
                 .flags = static_cast<UpdateFlags>(UpdateFlags::F_IS_SNAPSHOT | (it_update == it_last ? UpdateFlags::F_END_EVENT : UpdateFlags::F_NONE)),
-                .side = side
-            });
+                .side = side});
         }
 
-        if (update.event_time > symbol->order_book_->lastUpdateTime()) {
+        if (update.event_time > symbol->order_book_->lastUpdateTime())
+        {
             symbol->order_book_->setLastUpdate(update.event_time, seq_num);
         }
     }
 
-    if (it != pending_l2_subscriptions.end()) {
+    if (it != pending_l2_subscriptions.end())
+    {
         symbol->order_book_->populateL2SubscriptionResponse(md_queue_, coinbase::WebSocketChannel::LEVEL2);
         pending_l2_subscriptions.erase(it);
         symbol->md_level_update_cache_.clear();
         // TODO: publish MD order update
         symbol->md_order_update_cache_.clear();
     }
-    else {
+    else
+    {
         symbol->order_book_->populateL2Snapshot(md_queue_);
         symbol->md_level_update_cache_.clear();
         // TODO: publish MD order update
@@ -146,22 +170,25 @@ void CoinbaseExchange::onLevel2Snapshot(WebSocketClient* /* client */, uint64_t 
     LOG_DEBUG(symbol->order_book_->to_string());
 }
 
-void CoinbaseExchange::onLevel2Updates(WebSocketClient* /* client */, uint64_t seq_num, const coinbase::Level2UpdateBatch& update_batch) {
+void CoinbaseExchange::onLevel2Updates(WebSocketClient * /* client */, uint64_t seq_num, const coinbase::Level2UpdateBatch &update_batch)
+{
     auto *symbol = sym_mgr.getSymbol(update_batch.product_id);
-    if (!symbol) [[unlikely]] {
+    if (!symbol) [[unlikely]]
+    {
         return;
     }
 
     // Get or create event state for this symbol
-    auto& state = symbol_event_state_[symbol];
+    auto &state = symbol_event_state_[symbol];
 
     // Add each update to per-symbol pending_events queue
-    for (auto it = update_batch.updates.rbegin(), it_last = std::prev(update_batch.updates.rend()); it != update_batch.updates.rend(); ++it) {
+    for (auto it = update_batch.updates.rbegin(), it_last = std::prev(update_batch.updates.rend()); it != update_batch.updates.rend(); ++it)
+    {
         auto &update = *it;
         Event evt;
         evt.type = EventType::LEVEL_UPDATE;
-        evt.event_time = update.event_time;  // Use Coinbase timestamp
-        evt.sequence_id = state.next_sequence_id++;  // Internal counter
+        evt.event_time = update.event_time;         // Use Coinbase timestamp
+        evt.sequence_id = state.next_sequence_id++; // Internal counter
         evt.seq_num = seq_num;
         evt.received_time = utils::get_current_time_ns();
         evt.price = to_price_t(update.price_level);
@@ -170,30 +197,35 @@ void CoinbaseExchange::onLevel2Updates(WebSocketClient* /* client */, uint64_t s
         evt.flags = it == it_last ? UpdateFlags::F_END_EVENT : UpdateFlags::F_NONE;
 
         // Update last_event_time to track latest event_time seen
-        if (state.last_event_time < evt.event_time) {
+        if (state.last_event_time < evt.event_time)
+        {
             state.last_event_time = evt.event_time;
         }
 
         // LOG_DEBUG("Received L2 update for symbol {}: price={}, qty={}, side={}, event_time={}, seq_num={}, flags={}",
         //     symbol->symbol_, evt.price, evt.qty, evt.side, evt.event_time, evt.seq_num, evt.flags);
-        
+
         // Add to per-symbol priority queue
         state.pending_events.push(std::move(evt));
     }
 }
 
-void CoinbaseExchange::onMarketTradesSnapshot(WebSocketClient* /* client */, uint64_t seq_num, const std::vector<coinbase::MarketTrade>& snapshots) {
+void CoinbaseExchange::onMarketTradesSnapshot(WebSocketClient * /* client */, uint64_t seq_num, const std::vector<coinbase::MarketTrade> &snapshots)
+{
     auto &pending_subscriptions = pending_md_subscription_[coinbase::WebSocketChannel::MARKET_TRADES];
-    if (pending_subscriptions.empty() || snapshots.empty()) {
+    if (pending_subscriptions.empty() || snapshots.empty())
+    {
         return;
     }
-    Symbol* symbol = sym_mgr.getSymbol(snapshots[0].product_id);
-    if (!symbol || !pending_subscriptions.contains(symbol)) {
+    Symbol *symbol = sym_mgr.getSymbol(snapshots[0].product_id);
+    if (!symbol || !pending_subscriptions.contains(symbol))
+    {
         return;
     }
 
     auto it = trade_snapshots_.find(symbol);
-    if (it == trade_snapshots_.end()) {
+    if (it == trade_snapshots_.end())
+    {
         it = trade_snapshots_.emplace(symbol, utils::RingBuffer<Event>(16)).first;
     }
 
@@ -201,24 +233,25 @@ void CoinbaseExchange::onMarketTradesSnapshot(WebSocketClient* /* client */, uin
 
     auto sz = static_cast<uint32_t>(sizeof(MarketDataUpdate) + sizeof(MDSubscriptionResponse) + sizeof(MDTradeUpdate) + snapshots.size() * sizeof(MDTrade));
     auto index = md_queue_.reserve(sz);
-    auto *update = reinterpret_cast<MarketDataUpdate*>(md_queue_[index]);
+    auto *update = reinterpret_cast<MarketDataUpdate *>(md_queue_[index]);
     memcpy(update->symbol, symbol->symbol_.c_str(), sizeof(update->symbol));
     update->venue = venue_;
     update->type = MDUpdateType::SUB_RESPONSE;
 
-    auto *response = reinterpret_cast<MDSubscriptionResponse*>(update->data);
+    auto *response = reinterpret_cast<MDSubscriptionResponse *>(update->data);
     response->channel = coinbase::WebSocketChannel::MARKET_TRADES;
 
-    auto *snapshot = reinterpret_cast<MDTradeUpdate*>(response->data);
+    auto *snapshot = reinterpret_cast<MDTradeUpdate *>(response->data);
     snapshot->num_trades = static_cast<uint32_t>(snapshots.size());
 
-    auto *trades = reinterpret_cast<MDTrade*>(snapshot->trades);
+    auto *trades = reinterpret_cast<MDTrade *>(snapshot->trades);
 
-    auto& state = symbol_event_state_[symbol];
+    auto &state = symbol_event_state_[symbol];
 
     // Coinbase trades are in reverse chronological order
-    for (auto i = snapshots.size() - 1; i >= 0; i--) {
-        const auto& trade = snapshots[i];
+    for (auto i = static_cast<int32_t>(snapshots.size()) - 1; i >= 0; i--)
+    {
+        const auto &trade = snapshots[i];
         trades->price = to_price_t(trade.price);
         trades->qty = to_qty_t(trade.size);
         trades->side = static_cast<Side>(trade.side);
@@ -237,7 +270,8 @@ void CoinbaseExchange::onMarketTradesSnapshot(WebSocketClient* /* client */, uin
 
         trades++;
     }
-    for (const auto& trade : snapshots) {
+    for (const auto &trade : snapshots)
+    {
         trades->price = to_price_t(trade.price);
         trades->qty = to_qty_t(trade.size);
         trades->side = static_cast<Side>(trade.side);
@@ -247,28 +281,32 @@ void CoinbaseExchange::onMarketTradesSnapshot(WebSocketClient* /* client */, uin
     pending_subscriptions.erase(symbol);
 }
 
-void CoinbaseExchange::onMarketTrades(WebSocketClient* /* client */, uint64_t seq_num, const std::vector<coinbase::MarketTrade>& trades) {
-    for (const auto& trade : trades) {
+void CoinbaseExchange::onMarketTrades(WebSocketClient * /* client */, uint64_t seq_num, const std::vector<coinbase::MarketTrade> &trades)
+{
+    for (const auto &trade : trades)
+    {
         auto *symbol = sym_mgr.getSymbol(trade.product_id);
-        if (!symbol) [[unlikely]] {
+        if (!symbol) [[unlikely]]
+        {
             continue;
         }
 
         // Get or create event state for this symbol
-        auto& state = symbol_event_state_[symbol];
+        auto &state = symbol_event_state_[symbol];
 
         Event evt;
         evt.type = EventType::TRADE;
         evt.event_time = trade.time;
         evt.seq_num = seq_num;
-        evt.sequence_id = state.next_sequence_id++;  // Internal counter
+        evt.sequence_id = state.next_sequence_id++; // Internal counter
         evt.received_time = utils::get_current_time_ns();
         evt.price = to_price_t(trade.price);
         evt.qty = to_qty_t(trade.size);
         evt.side = static_cast<Side>(trade.side);
 
         // Update last_event_time to track latest event_time seen
-        if (state.last_event_time < evt.event_time) {
+        if (state.last_event_time < evt.event_time)
+        {
             state.last_event_time = evt.event_time;
         }
 
@@ -277,16 +315,20 @@ void CoinbaseExchange::onMarketTrades(WebSocketClient* /* client */, uint64_t se
     }
 }
 
-void CoinbaseExchange::onMarketDataGap(WebSocketClient* /* client */) {
+void CoinbaseExchange::onMarketDataGap(WebSocketClient * /* client */)
+{
     LOG_WARN("CoinbaseExchange Market Data Gap detected");
 }
-void CoinbaseExchange::onUserDataGap(WebSocketClient* /* client */) {
+void CoinbaseExchange::onUserDataGap(WebSocketClient * /* client */)
+{
     LOG_WARN("CoinbaseExchange User Data Gap detected");
 }
-void CoinbaseExchange::onMarketDataError(WebSocketClient* client, std::string &&err) {
-    LOG_ERROR("Websocket {:p} onMarketDataError {}", (void*)client, std::move(err));
+void CoinbaseExchange::onMarketDataError(WebSocketClient *client, std::string &&err)
+{
+    LOG_ERROR("Websocket {:p} onMarketDataError {}", (void *)client, std::move(err));
 }
 
-void CoinbaseExchange::onUserDataError(WebSocketClient* client, std::string &&err) {
-    LOG_ERROR("Websocket {:p} onUserDataError {}", (void*)client, std::move(err));
+void CoinbaseExchange::onUserDataError(WebSocketClient *client, std::string &&err)
+{
+    LOG_ERROR("Websocket {:p} onUserDataError {}", (void *)client, std::move(err));
 }
